fix test failure for windows
